+++ src/GeraNESApp/SDLAudioOutput.h
@@ -19,7 +19,7 @@ private:
 
     SDL_AudioSpec spec{};
 
-    uint32_t sampleAcc = 0;
+    double sampleAcc = 0.0;
     float m_volume = 1.0f;
     float m_sampleFormatScale = 256.0f;
